2022-07-26 Tech Meeting
This is the model I proposed at the end of the meeting. Notable changes:
* Added the missing inherits from `Collection` to `Element`.
* Renamed `Document` to `SpdxDocument`.
* Added `Bundle` and moved `BOM` to inherit from `Bundle` to remove the implication that `BOM`'s need to be a serialization root.
* Moved `SpdxDocument` to inherit from `Bundle` since they are a special type of bundle.

There is an open question on whether the distinction between `SpdxDocument` and `Bundle` is necessary, but this proposal doesn't force you to use `SpdxDocument` while providing a clear migration path from SPDX 2.x.

If during the formal definition of the classes it becomes clear we don't need both they can be collapsed together, but for now this would unblock us.