Create sbomType.md (#229)
* Create sbomType.md

With the publication of https://www.cisa.gov/sites/default/files/2023-04/sbom-types-document-508c.pdf on April 21, 2023.  We have agreed on standard definitions of the types of SBOMs.   Adding this property to the SBOM class so this information can be shared between tools.

* Update sbomType.md

make consistent SBOMType property with Vocabulary

* Update Sbom.md

Add SBOM types as optional property in the Sbom Class

Signed-off-by: Kate Stewart <kstewart@linuxfoundation.org>
Co-authored-by: Armin Tänzer <113098944+armintaenzertng@users.noreply.github.com>