Create ProfileIdentifierType.md (#249)
* Create ProfileIdentifierType.md

An arbitrary string for profile identifier, is going to cause too many snags, with arbitrary matching and capitalization.     Recommend restricting this down to the agreed on set of formal names for profiles.

* Update profile.md

Change range to be a ProfileIdentifierType rather than an XSD string type.

* Delete ProfileIdentifier.md

Make the profile identifier a set of enumerated values, rather than arbitrary string.    Added a Vocabulary of ProfileIdentifierType and remove the class which references a string type.