Create standard.md (#290)
* Create standard.md

Standards are being captured today in SPDX 2.3 in a comment field, and the AI, Dataset, and Usage profile all want to capture them as well.   Moving this as an optional field on artifact, so that there is a standard way to capture this relevant information.

* Update standard.md

* Update Artifact.md

Add optional standard that can be affiliated with artifact.