Remove namespaces from the logical model
Partial solution to #390 by removing the namespace and NamespaceMap
entirely from the model.  Note that this was suggested as a separate
PR and comment in PR #403 which still kept the NamespaceMap for
SPDX Documents.  Once this PR is agreed to / merged, we can create
a separate PR to add a namespace map hint to SPDX documents.

Signed-off-by: Gary O'Neall <gary@sourceauditor.com>