Proposal for Lite using profile restrictions
This is an initial proposal for how we may describe the Lite profile
and serialization.

The profile describes the additional restrictions and the serialization
describe the subset of supported fields.

This is just a draft at this point and does not include the complete
set of classes and properties for the profile and the serialization.

Signed-off-by: Gary O'Neall <gary@sourceauditor.com>