Clarify RelationshipTypes (#548)
The current wording of RelationshipTypes is very difficult to understand
and use correctly. It is very difficult to determine the directionality
of relationships, and often the documentation is difficult to parse as
its inverted from the expected directionality.

The renames most relationship types to complete the sentence:
`from` (is) (a) `RELATIONSHIP` `to`

Also, as much as possible the documentation is worded so that `from`
comes before `to` to further reduce confusion

All relationships are renamed assuming that the text description is the
authoritative description of the direction, not the name of the
relationship itself. This means existing implementation should be able
to find and replace with the new names and still be correct (assuming
they were able to figure them out in the first place).

For the sake of easier diffing, the relationships are not sorted based
on their new names.

Many of the security relationship were not clearly documented enough to
determine the correct directionality and still need to be addressed

Props to AI & Dataset for giving clear relationships that didn't need
modified.

Closes #534
Closes #533