Cleanup relationships (#556)
* Clarify security RelationshipTypes

This commit clarifies the directionality of security relationship types.

Works towards #533
Works towards #534

Signed-off-by: Rose Judge <rjudge@vmware.com>

* RelationshipType: Flip ancestor relationship

Flips the directionionality of the ancestor relationship from
`hasAncestor` to `ancestorOf`. A high fan-out makes more sense for this
relationship as a single parent Element is more likely to have multiple
children than a single child to have multiple parents.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

* RelationshipType: Combine Tool relationships

The tool relationships can be combined to a LifecycleScopedRelationship

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

* RelationshipType: Remove dependency relationships

Removes dependency relationships that can be covered by dependsOn with a
a LifecycleScopeType period

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

* RelationshipType: Sort relationships

Sorts the relationships by name. They were un-sorted due to being
renamed with better descriptions, but since this is complete the list
should remain sorted.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

* RelationshipType: Rename configOf -> configures

This wording is more clear to users about hos the relationship should be
used

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

* RelationshipType: Add Lifecycle Scopes to several relationships

These relationships can all occur during multiple points in the
lifecycle, so add the clarification to the documentation.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

* RelationshipType: invokedBy: Remove Build specific description

Since `invokedBy` is LifecycleScoped, there is no reason it can only
relate build elements

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

* RelationshipType: Clarify packagedBy description

Change "packaged form" -> "packaged instances" as this is more clear

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

* RelationshipType: Clarify hasOptionalDependency

There is no reason for the explanation to be backward for this
relationship, so switch it around.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

* RelationshipType: Rename hasMetafile -> hasMetadata

Metadata is more generic since the data may not live in a file

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>

---------

Signed-off-by: Rose Judge <rjudge@vmware.com>
Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>
Co-authored-by: Rose Judge <rjudge@vmware.com>