Move AnyLicenseInfo to Core
Moves the AnyLicenseInfo class to core. This makes sense so that
specific profiles can implement specific licensing implementations, but
other profiles do not need to be aware of them and can use the
AnyLicenseInfo from core to refer to a license without any specific
implementation.

As an example, SpdxDocument from Core references AnyLicenseInfo, so it
shouldn't live in SimpleLicense profile.