Remove the SoftwareDependencyRelationship Class and it's related properties.  (#585)
* Delete model/Software/Classes/SoftwareDependencyRelationship.md

As per discussion in Tech call in November,  remove the SoftwareDependency class for now, as it isn't being used with relationships.   Reconsider introducing for 3.1

Signed-off-by: Kate Stewart <kate.stewart@att.net>

* Delete model/Software/Properties/softwareLinkage.md

Remove property, as class using it is removed. 

Signed-off-by:  Kate Stewart <kate.stewart@att.net>

* Delete model/Software/Properties/conditionality.md

Remove property, as the class using it is removed as well. 

Signed-off-by:  Kate Stewart <kate.stewart@att.net>

* Delete model/Software/Vocabularies/SoftwareDependencyLinkType.md

Remove the related vocabularies.

* Delete model/Software/Vocabularies/DependencyConditionalityType.md

Remove unused vocabulary type.

---------

Signed-off-by: Kate Stewart <kate.stewart@att.net>