Remove hash algorithms with no description
The hash algorithms have only "TODO" for comments.
I could not find any record of these in the minutes and no one
has volunteered to fill in the description, so this PR deletes
these algorithms from the model.

If anyone would like to add them back, please add them with a
description as a separate pull request.

Signed-off-by: Gary O'Neall <gary@sourceauditor.com>