Fix delegatedTo Relationship
In practice, the definition of the delegatedTo relationship was flawed.
Since it linked one agent to another, it was impossible to determine
which delegatedTo relationships applied to which invokedBy
relationships.

To solve this, reword the delegatedTo relationship to indicated that the
`to` end must point to the specific invokedBy relationships to which it
applies.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>