Create Support Level to associate different types of support with an artifact.  (#628)
* Create SupportType.md

The support available for a software artifact is a factor increasingly being used in risk assessment.  equirement of listing the support type available.   This set of enumerations has been based on https://healthsectorcouncil.org/wp-content/uploads/2023/03/Health-Industry-Cybersecurity-Managing-Legacy-Technology-Security-HIC-MaLTS.pdf, with additional categories being added after discussion with the AI & Data profile working groups.   It should be able to be applied to software components, AI/ML trained models and datasets.

* Create supportLevel

Add information about the level of support that can be expected for a specific artifact.

* Update Artifact.md

Add the supportLevel property to the artifact

Signed-off-by:  Kate Stewart <kstewart@linuxfoundation.org>