Add Referenceable state
Adds metadata to indicate how classes can be referenced in serialized
documents:

"no" Means the class can only be referenced in one place (e.g. be the
RDF object), and can only have a blank node ID. It should be serialized
inline in serialization formats that support it

"local" means the class can be referenced multiple times, but only with
a blank node

"yes" means the class must have an external URI ID (e.g. not a blank
node), but the class itself may be inlined.