Reference License Expressions annex and cleanup
The primary change in this commit is to reference the SPDX License
Expressions annex (currently Annex D) in the LicenseExpression-
related class and property definitions.

This is to avoid repeated text, and also to make it clear that
license expression strings must conform to the license expression
syntax in order to be valid.

This commit also fixes a couple of minor typos in the prior
descriptions in SimpleLicensing.

Signed-off-by: Steve Winslow <steve@swinslow.net>