"trained by" -> "trained on"
Change from

"The `from` Element has been trained **by** the `to` Element(s)"

to

"The `from` Element has been trained **on** the `to` Element(s)"

consistent to the name of the type.

"testedOn" type also use "tested on"

"by" may signify that the `to` is an Agent, which is not the case here.

Signed-off-by: Arthit Suriyawongkul <arthit@gmail.com>