simplelicensing: Change custom license additions to link to an Element
Using a DictionaryEntry to map license additions in the `customIdToUri`
field has the unfortunate side effect that the dictionary entry value
will not be resolve to an actual object by either SHACL or code
bindings, since they are unaware that the value is supposed to be an
object IRI.

As such, create a new `ElementMap` object which can be used to map
string keys to Elements, and use this in a new `customIdToLicense`
property. The original `customIdToUri` property is retained for backward
compatibility, but marked as deprecated.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>